+++ Src/ITWeiHan.HtmlTableHelper.Test/CustomAttributeTests.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Collections.Generic;
 using System.Linq;
 
 namespace HtmlTableHelper.Test
@@ -8,14 +7,23 @@ namespace HtmlTableHelper.Test
     public class CustomAttributeTests
     {
         #region TestData
-        public class Test
+        public class ClassWithAttr
+        {
+            [TableColumn(DisplayName = "Column1", Skip = false)]
+            public string MyProperty1 { get; set; }
+            [TableColumn(DisplayName = "Column2")]
+            public string MyProperty2 { get; set; }
+        }
+
+        public class ClassWithSkipAttr
         {
             [TableColumn(DisplayName = "Column1", Skip = true)]
             public string MyProperty1 { get; set; }
             [TableColumn(DisplayName = "Column2", Skip = false)]
             public string MyProperty2 { get; set; }
         }
-        public class TestWithoutSkip
+
+        public class ClassWithoutSkip
         {
             [TableColumn(DisplayName = "Column1")]
             public string MyProperty1 { get; set; }
@@ -28,9 +36,9 @@ namespace HtmlTableHelper.Test
         public void Display_Test()
         {
             var expected = @"<table><thead><tr><th>Column1</th><th>Column2</th></tr></thead><tbody><tr><td>MyProperty1Value1</td><td>MyProperty1Value2</td></tr><tr><td>MyProperty2Value1</td><td>MyProperty2Value2</td></tr></tbody></table>";
-            var sourceData = new Test[] {
-                new Test { MyProperty1 = "MyProperty1Value1", MyProperty2 = "MyProperty1Value2" },
-                new Test { MyProperty1 = "MyProperty2Value1", MyProperty2 = "MyProperty2Value2" },
+            var sourceData = new ClassWithAttr[] {
+                new ClassWithAttr { MyProperty1 = "MyProperty1Value1", MyProperty2 = "MyProperty1Value2" },
+                new ClassWithAttr { MyProperty1 = "MyProperty2Value1", MyProperty2 = "MyProperty2Value2" },
             };
 
             var html = sourceData.ToHtmlTable();
@@ -41,9 +49,9 @@ namespace HtmlTableHelper.Test
         public void Skip_Attribute_Test()
         {
             var excepted = "<table><thead><tr><th>Column2</th></tr></thead><tbody><tr><td>MyProperty1Value2</td></tr><tr><td>MyProperty2Value2</td></tr></tbody></table>";
-            var sourceData = new Test[]{
-                new Test { MyProperty1 = "MyProperty1Value1", MyProperty2 = "MyProperty1Value2" },
-                new Test { MyProperty1 = "MyProperty2Value1", MyProperty2 = "MyProperty2Value2" },
+            var sourceData = new ClassWithSkipAttr[]{
+                new ClassWithSkipAttr { MyProperty1 = "MyProperty1Value1", MyProperty2 = "MyProperty1Value2" },
+                new ClassWithSkipAttr { MyProperty1 = "MyProperty2Value1", MyProperty2 = "MyProperty2Value2" },
             };
             var html = sourceData.ToHtmlTable();
             Assert.AreEqual(excepted, html);
@@ -53,9 +61,9 @@ namespace HtmlTableHelper.Test
         public void AllPorpsAreWithOutSkipTrue_Test()
         {
             var excepted = "<table><thead><tr><th>Column1</th><th>Column2</th></tr></thead><tbody><tr><td>MyProperty1Value1</td><td>MyProperty1Value2</td></tr><tr><td>MyProperty2Value1</td><td>MyProperty2Value2</td></tr></tbody></table>";
-            var sourceData = new TestWithoutSkip[]{
-                new TestWithoutSkip { MyProperty1 = "MyProperty1Value1", MyProperty2 = "MyProperty1Value2" },
-                new TestWithoutSkip { MyProperty1 = "MyProperty2Value1", MyProperty2 = "MyProperty2Value2" },
+            var sourceData = new ClassWithoutSkip[]{
+                new ClassWithoutSkip { MyProperty1 = "MyProperty1Value1", MyProperty2 = "MyProperty1Value2" },
+                new ClassWithoutSkip { MyProperty1 = "MyProperty2Value1", MyProperty2 = "MyProperty2Value2" },
             };
             var html = sourceData.ToHtmlTable();
             Assert.AreEqual(excepted, html);
@@ -65,11 +73,11 @@ namespace HtmlTableHelper.Test
         public void GetCustomAttribute_Test()
         {
             var excepted = new[]{
-                new TableColumnAttribute() { DisplayName = "Column1", Skip = true },
+                new TableColumnAttribute() { DisplayName = "Column1", Skip = false },
                 new TableColumnAttribute() { DisplayName = "Column2", Skip = false }
             };
 
-            var customAttributes = HtmlTableHelper.CustomAttributeHelper.GetCustomAttributes(type: typeof(Test));
+            var customAttributes = HtmlTableHelper.CustomAttributeHelper.GetCustomAttributes(type: typeof(ClassWithAttr));
 
             customAttributes.Select((attribute, index) =>
             {
@@ -84,13 +92,13 @@ namespace HtmlTableHelper.Test
         [TestMethod]
         public void GetCustomAttribute_RenderHtmlTable_Test()
         {
-            
+
             var exceptedHtml = @"<table><tr><th>Column1<th><th>Column2<th></tr><tr><td>Row1Column1</td><td>Row1Column2</td></tr><tr><td>Row2Column1</td><td>Row2Column2</td></tr></table>";
-            var sourceData = new Test[] { new Test { MyProperty1 = "Row1Column1", MyProperty2 = "Row1Column2" },
-                new Test { MyProperty1 = "Row2Column1", MyProperty2 = "Row2Column2" }
+            var sourceData = new ClassWithAttr[] { new ClassWithAttr { MyProperty1 = "Row1Column1", MyProperty2 = "Row1Column2" },
+                new ClassWithAttr { MyProperty1 = "Row2Column1", MyProperty2 = "Row2Column2" }
             };
 
-            var customAttributes = HtmlTableHelper.CustomAttributeHelper.GetCustomAttributes(type: typeof(Test));
+            var customAttributes = HtmlTableHelper.CustomAttributeHelper.GetCustomAttributes(type: typeof(ClassWithAttr));
 
             var th = string.Join(string.Empty, customAttributes.Select(s => $"<th>{s.DisplayName}<th>"));
             var td = string.Join(string.Empty, sourceData.Select(s => $"<tr><td>{s.MyProperty1}</td><td>{s.MyProperty2}</td></tr>"));