+++ Changelog.md
@@ -1,5 +1,5 @@
 ### version 1.1.6
-- [X] Support Annotation DisplayName/Filiter/Custom Attribute
+- [X] Support Annotation DisplayName/Skip Custom Attribute
 
 ### version 1.1.5
 - [X] Fix Cache Key GetHashCode replaced by type.MetadataToken + type.Module.ModuleVersionId + prop.MetadataToken + prop.Module.ModuleVersionId